Fixed UI. Added Missile Launchers
I fixed a bit with the UI. I did't like how stretched the UI got with the boss bg and border. So I adjusted It in Piskel to make It somewhat better imo.

For the Missile Launchers, they are now correctly made in the project. The transitions could be wonky, but not needed to have imo.

But the Idle versions should be working.

Oh and since I made some changes to the UI I have worked on the level.