UI + Sprites + Gameplay
UI: I fixed the Boss UI for the battle. I also changed some of the colors and placements a tiny bit. Also I added 2 versions of yours UI BG & Border for the UI.

Sprites: I added the Missile Launchers for 100% and 50% health.

Gaameplay: I changed the BlueFire Damage from 50 to 20. It should take a bit longer to destroy the boss. ani for laser is now 0.5 to 0.6.