Level Fixed 2/3
I am more or less happy with how the level is right now. I didn't add too many things but I wanted to make the level look a bit better.

I think the boss is better now as well. But I would request a playtest for the boss battle before we go with the green lights.

I still have 2 things I want to add before we are done though.